Create Temperature Converter
Temperature Converter" program is a program that prompts the user to enter a temperature in Celsius or Fahrenheit and then convert it to the other unit of measurement.

Here's a brief explanation of how this program can be implemented:
->First, the program will prompt the user to enter a temperature and the unit of measurement (Celsius or Fahrenheit).
->Next, the program will convert the temperature to the other unit of measurement using the appropriate conversion formula.
->Then program will then display the converted temperature.


here 'scanf()' function is used to read the user input, and if and else if statements are used to check the unit of measurement entered by the user. The appropriate conversion formula is used to convert the temperature, and the result is displayed using 'printf()'.