Create GUESS THE NUMBER
 "Guess the Number" game. This game will generate a random number and prompt the user to guess the number. The program will provide feedback to the user after each guess, indicating whether the guess was too high or too low.

Here's a brief explanation of how this program can be implemented:
->First, the program will generate a random number between a specified range using the java.util.Random class.
->Next, the program will prompt the user to guess the number.
->The program will read the user's guess using the java.util.Scanner class.
->The program will compare the user's guess to the random number and provide feedback indicating whether the guess was too high, too low, or correct.
->If the guess was incorrect, the program will repeat steps 2-4 until the user guesses the correct number.

In this program, the java.util.Scanner class is used to read user input, and the java.util.Random class is used to generate a random number. The do-while loop is used to repeatedly prompt the user for guesses until the correct number is guessed. Feedback is provided to the user using if-else statements.